requirements.txt
# GUI Framework (PySide6 >= 6.5 bundles WebEngine)
PySide6>=6.6.0

# Generator dependencies
jinja2>=3.1.2
markdown>=3.5.1
pyyaml>=6.0.1
ebooklib>=0.18
beautifulsoup4>=4.12.2
Pillow>=10.1.0
watchdog>=3.0.0
websockets>=12.0
htmlmin>=0.1.12
rcssmin>=1.1.1
rjsmin>=1.2.0

# Gradio Studio (optional)
gradio>=4.0.0
